+++ assignment2_codes/assignment2_codes/finDif/aux/GridConstructor.h
@@ -2,7 +2,7 @@
 //  GridConstructor.h
 //  ComputationalFinance
 //
-//  Created by Rubén Fernández Fuertes on 15/4/21.
+//  Created on 15/4/21.
 //
 
 #include <vector>
@@ -12,7 +12,7 @@ using namespace std;
 
 class CGridConstructor{
 public:
-    static vector<double>   setupStockPrices(double dS, int J);
+    static vector<double>   setupStockPrices(double dS, long J);
     static double           boundaryConditions(int j);
     static double           lagrangeInterpolation(const vector<double>& y,const vector<double>& x, double x0, unsigned int n);
 };
